Settings Window
Settings allow the user to customize the user interface to individual
needs and preferences.
The features added in this branch include:

Settings Window
Standing Pieces can be placed diagonally
Piece size of flat and standing stones can be scaled
Antialiasing-support
The automatic show/hide of the chat window can be disabled.