Check for wins in a loaded TPS
Previous code for loading a TPS did not check whether the loaded
position was a win. Now we check for all possible wins.

Added an enhancement to the gameover() method so that the caller could
optionally provide a bit more information to the output message.

Also, fixed a small defect in the original code in which a bad socket
connection tried to invoke "this.output()" but should have just been
"output()".